refactor gam cross validation
