+++ statsmodels/sandbox/gam_gsoc2015/gam_cross_validation/__init__.py
@@ -1,0 +1,1 @@
+__author__ = 'benny'
